+++ spec/formatter/exception_formatter_spec.rb
@@ -7,7 +7,7 @@ describe Lumberjack::Formatter::ExceptionFormatter do
     formatter = Lumberjack::Formatter::ExceptionFormatter.new
     expect(formatter.call(e)).to eq("ArgumentError: not expected")
   end
-  
+
   it "should convert an exception with a backtrace to a string" do
     begin
       raise ArgumentError.new("not expected")
@@ -16,5 +16,15 @@ describe Lumberjack::Formatter::ExceptionFormatter do
       expect(formatter.call(e)).to eq("ArgumentError: not expected#{Lumberjack::LINE_SEPARATOR}  #{e.backtrace.join(Lumberjack::LINE_SEPARATOR + '  ')}")
     end
   end
-  
+
+  it "should clean the backtrace" do
+    begin
+      raise ArgumentError.new("not expected")
+    rescue => e
+      formatter = Lumberjack::Formatter::ExceptionFormatter.new
+      formatter.backtrace_cleaner = lambda { |lines| ["redacted: #{lines.size}"] }
+      expect(formatter.call(e)).to eq("ArgumentError: not expected#{Lumberjack::LINE_SEPARATOR}  redacted: #{e.backtrace.size}")
+    end
+  end
+
 end